requirements.txt
certifi==2020.6.20
Django==2.2
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2

djangorestframework-jwt~=1.11.0
bson~=0.5.10
pyyaml~=5.3.1